Lots of work on the dJacobi module.
The idea is that the dJacobi object builds the lightweight dEFS and
dRule objects which are actually used in the solution process.  There is
currently only one implementation (Tensor) which is probably the only
one for a while.  It is not finished, but getting there.

While polylib is workable, I would still like to renovate it a bit.
However, this is not a high priority so things will probably remain as
is for while.

There is still thinking to be done regarding element interface
conditions.  I have done nothing on this front so far, but dEFS can be
extended to store pointers to the interpolation operators (which may
from a different GaussFamily than the basis/deriv matrices).  The key
then is to change the calling sequence for dJacobiGetEFS() to include
the facet sizes as well as the primary basis size.

Signed-off-by: Jed Brown <jed@59A2.org>