Started the dFS object.
This object should handle constraining degrees of freedom and function
spaces on a subdomain of the mesh.  Eventually it will be a full-fledged
PETSc DM object.  For problems with mixed discretization, there will be
a dMFS (multi function space) object.  All fields in a dMFS will need to
be discretized over a dFS and that dFS must share the same dQuotient.
If this property is not satisfied, just use Barry's DMComposite.

Signed-off-by: Jed Brown <jed@59A2.org>