This is the first attempt at replacing the build system with CMake.
I'm mostly happy with this.  CMake syntax is pretty painful, but for
simple things it really does make life easy.

I think the FindMPI script is rather fragile.  There should be a way to
confirm that several libraries are configured to link against the same
MPI, for instance.  Somehow it is possible for a lib built against
mpich2 and linked with the correct (as far as I can tell) link line to
still have /usr/lib/libmpi.so (ompi).

It might be my fault, but we are linking in the wrong hdf5 in this case.
Certainly more work to be done here, but the builds are much nicer and
tests should be easier to automate.

Signed-off-by: Jed Brown <jed@59A2.org>