Add -dfs_assemble_reduced, uncouples components in matrix assembly.
This is particularly useful to improve AMG performance.  ML and
BoomerAMG don't have block matrix formats so inter-component coupling
significantly adds to the cost.  PETSc has block smoothers and a more
efficient storage scheme so these extra entries are normally very cheap
to leave in, even if the coupling is weak (leaving them in trades
metadata for useful data).  It would be really nice if the AMG packages
could just provide interpolation matrices, respecting the nodal
structure, and let PETSc handle the rest.