Fixed definition of ellip_p, add dirty LINEAR macro
Exponent should be (p-2)/2 instead of p-2 since gamma = (1/2)|Du|^2.
Updated forcing terms for exact solutions and tests.

For performance comparison, it is useful to just assemble the matrix as
if the problem is linear (to avoid computing 'pow' and other expensive
operations that are not needed in the linear case).  This is ugly and
should remain deactivated.