Check alignment in TensorMult_Hex_P4_Q4_D1
16-byte alignment is required for the MOVAPD instructions used
extensively in this function.  Prior to a fix in PetscTrMalloc, this
function was crashing with optimized PETSc and -malloc_dump.  The reason
for this: PetscTrMallocDefault() was only guaranteeing 8-byte alignment
(rather than respecting PETSC_MEMALIGN).  In debug mode, these happened
to be equivalent because struct _trSPACE contained a PetscStack member.
In optimized mode, the stack is not stored so we were only getting
8-byte alignment, causing MOVAPD to generate a GP fault.