Big header shuffle.
My intent is that dohp.h contains various macros that many users will
want, but is not mandatory.  In particular, it contains vararg macros
which cannot be seen from C++.  So dohptype.h contains necessary base
types and a user is only required to include dohpXXX.h to get whichever
interface they will use.

Also changed the headers to always use extern C.  I have no intent of
offering a name-mangled build of Dohp (like PETSc does), Dohp is a C99
library which can be called from C89 and C++, that is all.