A bit of renaming so that user code does not need PETSc headers unless it calls PETSc.
This required a couple forward declarations and reproducing a little logic
determining type sizes and real/complex.  All the C99 stuff should now be
isolated to dohp.h (which users are not required to include) and this is the
only place that PETSc headers other than petscconf.h are included.