Implement ValueCachePhysical for mapping the physical mapping between element- and patch-natural ordering
Note that both are required: dEFSApply() needs the mapping in
element-natural form, but the user needs the weights in patch-natural
form.

It is possible that dx/dX could be mapped to patch-natural ordering, and
then compute geometry, but that would change the dEFSApply() interface
so I don't want to do it now. Maybe it will make sense later.