Use deferred evaluation for manufactured solutions.
Speeds up code generation some, speeds up run-time by 30% or so for
elasticity and Stokes. Code generation and compilation could be further
sped up by noting that the physics part of the manufactured solutions
are now independent of the exact solution, therefore it would be enough
to compute the values, gradient, and Hessian for each exact solution,
but only do the physics part of the manufactured solution once.