The conservative transport term for energy is (E+p)*u, not (E+p)*ui
The previous form was correct for advection of internal energy, but the
volumetric term p*ui and transport of kinetic energy were incorrect. The
diffusive flux in the present formulation is obtained by considering moisture
transport in the reference frame of total velocity instead of ice velocity.
That is, a hydraulic gradient relative to the ice produces a transport term
but part of that term is already represented in transport of total energy
at the total velocity, therefore only the remaining component should be added.